lambda-connector-customizations/globals/connector/documents/requirements.txt
hasura-ndc==0.38
opentelemetry_sdk~=1.32.1
opentelemetry_api~=1.32.1
aiohttp~=3.11.14
